+++ sdk/search/search-documents/src/searchIndexerClient.ts
@@ -4,7 +4,7 @@
 import { isTokenCredential, KeyCredential, TokenCredential } from "@azure/core-auth";
 import { InternalClientPipelineOptions } from "@azure/core-client";
 import { ExtendedCommonClientOptions } from "@azure/core-http-compat";
-import { bearerTokenAuthenticationPolicy } from "@azure/core-rest-pipeline";
+import { bearerTokenAuthenticationPolicy, Pipeline } from "@azure/core-rest-pipeline";
 import { SearchIndexerStatus } from "./generated/service/models";
 import { SearchServiceClient as GeneratedClient } from "./generated/service/searchServiceClient";
 import { logger } from "./logger";
@@ -91,6 +91,11 @@ export class SearchIndexerClient {
   private readonly client: GeneratedClient;
 
   /**
+   * A reference to the internal HTTP pipeline for use with raw requests
+   */
+  public readonly pipeline: Pipeline;
+
+  /**
    * Creates an instance of SearchIndexerClient.
    *
    * Example usage:
@@ -139,6 +144,7 @@ export class SearchIndexerClient {
       this.serviceVersion,
       internalClientPipelineOptions,
     );
+    this.pipeline = this.client.pipeline;
 
     if (isTokenCredential(credential)) {
       const scope: string = options.audience